Demo Version
First demo i am happy to show.
added camera lock added a pivot point changer so the camera focuses on the center of the object added zoom in and zoom out based on the size of the window frame added a material selector to change between currently 2 materials but its done through blender so more materials can be added smoothly.